+++ spec/classy_enum/active_record_spec.rb
@@ -96,28 +96,12 @@ describe "A ClassyEnum that has a different field name than the enum" do
 end
 
 class ActiveDog < ActiveRecord::Base
-  classy_enum_attr :breed, :suffix => 'type'
   classy_enum_attr :color
-
-  validates :name,
-    :presence => true,
-    :uniqueness => { :scope => [:breed] }
-
   validates_uniqueness_of :name, :scope => :color
-
   scope :goldens, where(:breed => 'golden_retriever')
 end
 
 describe ActiveDog do
-
-  context 'valid instance' do
-    subject { ActiveDog.new(:name => 'sirius', :breed => :golden_retriever, :color => :black) }
-
-    it { should have(:no).errors_on(:breed) }
-    its(:breed_type) { should be_a_golden_retriever }
-    its(:breed) { should == 'golden_retriever' }
-  end
-
   context 'uniqueness on name' do
     subject { ActiveDog.new(:name => 'Kitteh', :breed => :golden_retriever, :color => :black) }
     it { should be_valid }
@@ -129,12 +113,6 @@ describe ActiveDog do
 
       it { should have(1).error_on(:name) }
     end
-  end
-
-  context 'invalid instance' do
-    subject { ActiveDog.new(:name => 'sirius', :breed => :golden_retrievers, :color => :white) }
-
-    it { should have(1).error_on(:breed) }
   end
 
   context 'scopes' do